+++ web/src/components/datasets/DatasetInfo.tsx
@@ -1,31 +1,19 @@
 // Copyright 2018-2024 contributors to the Marquez project
 // SPDX-License-Identifier: Apache-2.0
 import * as Redux from 'redux'
-import {
-  Box,
-  Card,
-  CardContent,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableRow,
-} from '@mui/material'
+import { Box, Table, TableBody, TableCell, TableHead, TableRow } from '@mui/material'
 import { Field, Run } from '../../types/api'
 import { IState } from '../../store/reducers'
 
 import { connect, useSelector } from 'react-redux'
 import { fetchJobFacets, resetFacets } from '../../store/actionCreators'
-import { stopWatchDuration } from '../../helpers/time'
 import Collapse from '@mui/material/Collapse'
 import DatasetTags from './DatasetTags'
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
-import MqCode from '../core/code/MqCode'
 import MqEmpty from '../core/empty/MqEmpty'
 import MqJsonView from '../core/json-view/MqJsonView'
 import MqText from '../core/text/MqText'
 import React, { FunctionComponent, useEffect, useState } from 'react'
-import RunStatus from '../jobs/RunStatus'
 
 export interface DispatchProps {
   fetchJobFacets: typeof fetchJobFacets
@@ -52,7 +40,7 @@ type DatasetInfoProps = {
   DispatchProps
 
 const DatasetInfo: FunctionComponent<DatasetInfoProps> = (props) => {
-  const { datasetFields, facets, run, jobFacets, fetchJobFacets, resetFacets } = props
+  const { datasetFields, facets, run, fetchJobFacets, resetFacets } = props
   const i18next = require('i18next')
   const dsNamespace = useSelector(
     (state: IState) => state.datasetVersions.result.versions[0].namespace
@@ -151,16 +139,14 @@ const DatasetInfo: FunctionComponent<DatasetInfoProps> = (props) => {
                     <TableRow>
                       <TableCell colSpan={4} style={{ padding: 0, border: 'none' }}>
                         <Collapse in={expandedRows.includes(index)} timeout='auto'>
-                          <Card>
-                            <CardContent>
-                              <DatasetTags
-                                namespace={dsNamespace}
-                                datasetName={dsName}
-                                datasetTags={field.tags}
-                                datasetField={field.name}
-                              />
-                            </CardContent>
-                          </Card>
+                          <Box p={2}>
+                            <DatasetTags
+                              namespace={dsNamespace}
+                              datasetName={dsName}
+                              datasetTags={field.tags}
+                              datasetField={field.name}
+                            />
+                          </Box>
                         </Collapse>
                       </TableCell>
                     </TableRow>
@@ -177,24 +163,6 @@ const DatasetInfo: FunctionComponent<DatasetInfoProps> = (props) => {
             <MqText subheading>{i18next.t('dataset_info.facets_subhead')}</MqText>
           </Box>
           <MqJsonView data={facets} aria-label={i18next.t('dataset_info.facets_subhead_aria')} />
-        </Box>
-      )}
-      {run && (
-        <Box mt={2}>
-          <Box mb={1}>
-            <Box display={'flex'} alignItems={'center'} justifyContent={'space-between'}>
-              <Box display={'flex'} alignItems={'center'}>
-                <RunStatus run={run} />
-                <MqText subheading>{i18next.t('dataset_info.run_subhead')}</MqText>
-              </Box>
-              <Box display={'flex'}>
-                <MqText bold>{i18next.t('dataset_info.duration')}&nbsp;</MqText>
-                <MqText subdued>{stopWatchDuration(run.durationMs)}</MqText>
-              </Box>
-            </Box>
-            <MqText subdued>{run.jobVersion && run.jobVersion.name}</MqText>
-          </Box>
-          {<MqCode code={(jobFacets?.sql as SqlFacet)?.query} language={'sql'} />}
         </Box>
       )}
     </Box>