+++ web/src/routes/column-level/ColumnLevelDrawer.tsx
@@ -14,7 +14,6 @@ import { IState } from '../../store/reducers'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import { fetchDataset } from '../../store/actionCreators'
-import { findConnectedNodes } from './layout'
 import { useSearchParams } from 'react-router-dom'
 import CloseIcon from '@mui/icons-material/Close'
 import IconButton from '@mui/material/IconButton'
@@ -53,9 +52,6 @@ const ColumnLevelDrawer = ({
     return null
   }
 
-  const column = searchParams.get('column')
-  const connectedColumns = findConnectedNodes(columnLineage.graph, column)
-
   return (
     <Box width={`${WIDTH}px`}>
       <Box p={2} display={'flex'} justifyContent={'space-between'} alignItems={'center'}>
@@ -76,65 +72,43 @@ const ColumnLevelDrawer = ({
       ) : (
         <>
           <Box p={2}>
-            <MqText subheading>Schema</MqText>
+            <MqText subheading>SCHEMA</MqText>
           </Box>
           {dataset.fields.length > 0 && (
-            <Table size='small'>
-              <TableHead>
-                <TableRow>
-                  <TableCell align='left'>
-                    <MqText subheading inline>
-                      {i18next.t('dataset_info_columns.name')}
-                    </MqText>
-                  </TableCell>
-                  <TableCell align='left'>
-                    <MqText subheading inline>
-                      {i18next.t('dataset_info_columns.type')}
-                    </MqText>
-                  </TableCell>
-                </TableRow>
-              </TableHead>
-              <TableBody>
-                {dataset.fields.map((field) => {
-                  return (
-                    <TableRow key={field.name}>
-                      <TableCell align='left'>{field.name}</TableCell>
-                      <TableCell align='left'>{field.type}</TableCell>
-                    </TableRow>
-                  )
-                })}
-              </TableBody>
-            </Table>
-          )}
-          {connectedColumns.length > 0 && column && (
             <>
-              <Box p={2}>
-                <MqText bold subheading>
-                  Connected Columns
-                </MqText>
-              </Box>
               <Table size='small'>
                 <TableHead>
                   <TableRow>
                     <TableCell align='left'>
                       <MqText subheading inline>
-                        Dataset
+                        {i18next.t('dataset_info_columns.name')}
                       </MqText>
                     </TableCell>
                     <TableCell align='left'>
                       <MqText subheading inline>
-                        Field
+                        {i18next.t('dataset_info_columns.type')}
                       </MqText>
                     </TableCell>
+                    <TableCell align='left'>
+                      <MqText subheading inline>
+                        {i18next.t('dataset_info_columns.description')}
+                      </MqText>
+                    </TableCell>
+                    <TableCell align='left'></TableCell>
                   </TableRow>
                 </TableHead>
                 <TableBody>
-                  {connectedColumns.map((c) => {
+                  {dataset.fields.map((field) => {
                     return (
-                      <TableRow key={c.id}>
-                        <TableCell align='left'>{c.data?.dataset || 'Unknown'}</TableCell>
-                        <TableCell align='left'>{c.data?.field || 'Unknown'}</TableCell>
-                      </TableRow>
+                      <React.Fragment key={field.name}>
+                        <TableRow>
+                          <TableCell align='left'>{field.name}</TableCell>
+                          <TableCell align='left'>{field.type}</TableCell>
+                          <TableCell align='left'>
+                            {field.description || 'no description'}
+                          </TableCell>
+                        </TableRow>
+                      </React.Fragment>
                     )
                   })}
                 </TableBody>
@@ -146,8 +120,7 @@ const ColumnLevelDrawer = ({
       {dataset && dataset.columnLineage && (
         <>
           <Box p={2}>
-            <MqText subheading>Column Lineage Facets</MqText>
-
+            <MqText subheading>FACETS</MqText>
             <MqJsonView data={dataset.columnLineage} />
           </Box>
         </>