+++ web/src/routes/table-level/TableLevel.tsx
@@ -108,6 +108,9 @@ const ColumnLevel: React.FC<ColumnLevelProps> = ({
           anchor={'right'}
           open={!!searchParams.get('tableLevelNode')}
           onClose={() => setSearchParams({})}
+          PaperProps={{
+            sx: { backgroundColor: theme.palette.background.default, backgroundImage: 'none' },
+          }}
         >
           <Box sx={{ pt: '98px' }}>
             <TableLevelDrawer />